requirements.txt
Flask==3.0.0
pymongo==4.6.1
python-dotenv==1.0.0
waitress==2.1.2
openai==1.6.1
jsonschema==4.21.1
